Added the ability for the user to book sessions, see his booked sesions and remove them if needed and refactored some functions

+++ fitly_backend/src/main/java/com/gymbooking/fitly/models/GymStatistics.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class GymStatistics {
     private int totalSessions;
     private int totalParticipants;
-    private int totalRevenue;
+    private double totalRevenue;
 }
