+++ fitly_backend/src/main/java/com/gymbooking/fitly/services/GymService.java
@@ -2,7 +2,6 @@ package com.gymbooking.fitly.services;
 
 import com.gymbooking.fitly.models.Gym;
 import com.gymbooking.fitly.models.GymStatistics;
-import com.gymbooking.fitly.models.Session;
 import com.gymbooking.fitly.models.User;
 import com.gymbooking.fitly.repositories.GymRepository;
 import com.gymbooking.fitly.repositories.UserRepository;
@@ -90,7 +89,7 @@ public class GymService {
     
         int totalSessions = gymRepository.countSessionsByGymId(gymId);
         int totalParticipants = gymRepository.countParticipantsByGymId(gymId);
-        int totalRevenue = gymRepository.calculateTotalRevenueByGymId(gymId);
+        double totalRevenue = gymRepository.calculateTotalRevenueByGymId(gymId);
         
         GymStatistics statistics = new GymStatistics();
         statistics.setTotalSessions(totalSessions);
